Moved meshing outside of disc and changed shared.py and test_base_disc
